Rename wayland_shell.{c,h} to shell.{c,h}
