+++ src/compositor/module.c
@@ -44,7 +44,7 @@
 #include "compositor/monitor_mode.h"
 #include "compositor/wayland/compositor.h"
 #include "compositor/wayland/seat.h"
-#include "compositor/wayland/wayland_shell.h"
+#include "compositor/wayland/shell.h"
 #include "logger/module.h"
 #include "util/cleaner.h"
 #include "util/wayland.h"
