+++ src/compositor/wayland/shell.c
@@ -29,7 +29,7 @@
 #include <wayland-server.h>
 
 #include "compositor/wayland/surface.h"
-#include "compositor/wayland/wayland_shell.h"
+#include "compositor/wayland/shell.h"
 #include "compositor/wayland/wayland_shell_surface.h"
 #include "util/arithmetical.h"
 #include "util/wayland.h"
